refactor. RacingCarGame의 이름을 controller패키지의 RacingCarGameController로 변경
